style: Problem4_3のdouble暗黙変換を明示的なキャストに修正する
Math.Pow の戻り値（double）を暗黙的に string 連結していたため、
(int) キャストを明示して整数として扱っていることを明確にする。

+++ src/Chapter04/Exercises.cs
@@ -30,7 +30,7 @@ public class Exercises
         string result = "";
         for (int i = 1; i <= 8; i++)
         {
-            result += Math.Pow(2, i);
+            result += (int)Math.Pow(2, i);
             if (i < 8) result += ",";
         }
         return result;
